requirements.txt
PySide6==6.8.1
PySide6-Fluent-Widgets
python-vlc
python-mpv

# Telegram integration
telethon>=1.36.0
qrcode>=7.0
pillow>=10.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Security
cryptography>=41.0.0
keyring>=24.0.0
tgcrypto>=1.2.0